refactor: migrate db tauri commands to drizzle
