+++ apps/desktop/src/lib/cmds/ext.ts
@@ -7,10 +7,11 @@ import { decideKkrpcSerialization } from "@/utils/kkrpc"
 import { sleep } from "@/utils/time"
 import { trimSlash } from "@/utils/url"
 import { constructExtensionSupportDir } from "@kksh/api"
-import { db, spawnExtensionFileServer } from "@kksh/api/commands"
+import { spawnExtensionFileServer } from "@kksh/api/commands"
 import type { HeadlessCommand } from "@kksh/api/headless"
 import { CustomUiCmd, ExtPackageJsonExtra, HeadlessCmd, TemplateUiCmd } from "@kksh/api/models"
 import { constructJarvisServerAPIWithPermissions, type IApp } from "@kksh/api/ui"
+import { db } from "@kksh/drizzle"
 import { launchNewExtWindow, loadExtensionManifestFromDisk } from "@kksh/extension"
 import type { IKunkunFullServerAPI } from "@kunkunapi/src/api/server"
 import { convertFileSrc } from "@tauri-apps/api/core"
