+++ apps/desktop/src/routes/app/extension/clipboard/+page.svelte
@@ -3,8 +3,9 @@
 	import { goHome } from "@/utils/route"
 	import { listenToNewClipboardItem, listenToWindowFocus } from "@/utils/tauri-events"
 	import Icon from "@iconify/svelte"
-	import { ClipboardContentType, db } from "@kksh/api/commands"
+	import { ClipboardContentType } from "@kksh/api/commands"
 	import { SearchModeEnum, SQLSortOrderEnum, type ExtData } from "@kksh/api/models"
+	import { db } from "@kksh/drizzle"
 	import { Button, Command, Resizable } from "@kksh/svelte5"
 	import { Constants } from "@kksh/ui"
 	import { CustomCommandInput, GlobalCommandPaletteFooter } from "@kksh/ui/main"
