+++ apps/desktop/src/routes/app/extension/ui-iframe/+page.ts
@@ -1,8 +1,7 @@
 import { KunkunIframeExtParams } from "@/cmds/ext"
 import { i18n } from "@/i18n"
-import { appState } from "@/stores/appState"
-import { db, unregisterExtensionWindow } from "@kksh/api/commands"
 import type { Ext as ExtInfoInDB, ExtPackageJsonExtra } from "@kksh/api/models"
+import { db } from "@kksh/drizzle"
 import { loadExtensionManifestFromDisk } from "@kksh/extension"
 import { error as svError } from "@sveltejs/kit"
 import { join } from "@tauri-apps/api/path"
