+++ apps/desktop/src/routes/app/extension/ui-worker/+page.ts
@@ -1,7 +1,7 @@
 import { KunkunTemplateExtParams } from "@/cmds/ext"
 import { i18n } from "@/i18n"
-import { db, unregisterExtensionWindow } from "@kksh/api/commands"
-import type { Ext as ExtInfoInDB, ExtPackageJsonExtra } from "@kksh/api/models"
+import type { ExtPackageJsonExtra } from "@kksh/api/models"
+import { db } from "@kksh/drizzle"
 import { loadExtensionManifestFromDisk } from "@kksh/extension"
 import { error as sbError, error as svError } from "@sveltejs/kit"
 import { join } from "@tauri-apps/api/path"
