+++ packages/api/src/commands/index.ts
@@ -5,8 +5,7 @@ export * from "./tools"
 export * from "./extension"
 export * from "./system"
 export * from "./store"
-export * as db from "./db"
-export { JarvisExtDB } from "./db"
+export * as sql from "./sql"
 export * from "./clipboard"
 export * from "./fileSearch"
 export * from "./utils"
