+++ packages/drizzle/index.ts
@@ -1,2 +1,3 @@
 export * as schema from "./drizzle/schema"
 export * as relations from "./drizzle/relations"
+export * as db from "./src/apis"
