+++ packages/extension/package.json
@@ -16,6 +16,7 @@
 	},
 	"dependencies": {
 		"@kksh/api": "workspace:*",
+		"@kksh/drizzle": "workspace:*",
 		"@std/semver": "npm:@jsr/std__semver@^1.0.4",
 		"@tauri-apps/plugin-upload": "^2.2.1",
 		"semver": "^7.7.1",
