+++ packages/extension/src/install.ts
@@ -3,8 +3,9 @@
  * including install, uninstall, upgrade, check app-extension compatibility, etc.
  */
 import { isCompatible } from "@kksh/api"
-import { copy_dir_all, db, decompressTarball } from "@kksh/api/commands"
+import { copy_dir_all, decompressTarball } from "@kksh/api/commands"
 import type { ExtensionStoreListItem, ExtPackageJsonExtra } from "@kksh/api/models"
+import { db } from "@kksh/drizzle"
 import { greaterThan, parse as parseSemver } from "@std/semver"
 import * as path from "@tauri-apps/api/path"
 import * as dialog from "@tauri-apps/plugin-dialog"
